Remove query module plugin logic; sql only
