+++ nerium/db.py
@@ -2,8 +2,21 @@
 """SQL database query implementation of resultset, using Records library
 to fetch a dataset from configured query_name
 """
+import os
 from sqlalchemy import create_engine
-from nerium.data_source import get_data_source
+
+
+def get_data_source(query):
+    """ Get data source connection metadata based on config. Prefer, in order:
+        - Query file frontmatter
+        - DATABASE_URL in environment
+    """
+    # from frontmatter
+    if "database_url" in query.metadata.keys():
+        return query.metadata["database_url"]
+
+    # Default to $DATABASE_URL with sqlite fallback if not set in frontmatter
+    return os.getenv("DATABASE_URL", "sqlite:///")
 
 
 def connection(query):
