+++ nerium/query.py
@@ -1,12 +1,12 @@
 import os
 from datetime import datetime
-from importlib import import_module
 from pathlib import Path
 from types import SimpleNamespace
 
 import sqlparse
 import yaml
 from jinja2.sandbox import SandboxedEnvironment
+from nerium.db import result
 
 
 def query_file(query_name):
@@ -34,7 +34,7 @@ def extract_metadata(query_string):
             ][0]
         )
     except IndexError:
-        return None
+        return {}
     meta_string = meta_comment.split("---")[1]
     metadata = yaml.safe_load(meta_string)
     return metadata
@@ -56,7 +56,6 @@ def parse_query_file(query_name):
 
         query_obj.metadata = extract_metadata(query_string)
         query_obj.path = query_path
-        query_obj.result_module = query_path.suffix.strip(".")
         query_obj.body = query_string
 
     except (FileNotFoundError, TypeError):
@@ -64,37 +63,6 @@ def parse_query_file(query_name):
         query_obj.status_code = 404
 
     return query_obj
-
-
-def plugin_module(name):
-    """Load all modules named like `nerium_*` as plugin registry and return
-    module matching query file stem if available, or default to sql built-in
-
-    Besides being named like `nerium_*`, plugins are expected to accept a query object
-    and provide a `result` method that returns an iterable dataset (list of dicts, e.g.)
-    """
-    from pkgutil import iter_modules
-
-    plugins = {
-        name: import_module(name)
-        for finder, name, ispkg in iter_modules()
-        if name.startswith("nerium_")
-    }
-    plugin_name = f"nerium_{name}"
-    if plugin_name in plugins.keys():
-        return plugins[plugin_name]
-    else:
-        return import_module("nerium.resultset.sql")
-
-
-def assign_module(module="sql"):
-    """Import resultset module matching query file suffix from nerium.resultset or plugin
-    """
-    try:
-        result_module = import_module(f"nerium.resultset.{module}")
-    except ModuleNotFoundError:
-        result_module = plugin_module(module)
-    return result_module
 
 
 def process_template(body, **kwargs):
@@ -115,11 +83,11 @@ def get_result_set(query_name, **kwargs):
     if query.error:
         return query
 
-    result_module = assign_module(query.result_module)
+    # result_module = assign_module(query.result_module)
 
     # TODO: New resultset object here instead of mutating query object
     query.body = process_template(body=query.body, **kwargs)
-    query.result = result_module.result(query, **kwargs)
+    query.result = result(query, **kwargs)
 
     # Set query.error in case result_module captures an excecption
     if isinstance(query.result[0], dict) and "error" in query.result[0].keys():
