--- tests/nerium_mock.py
@@ -1,2 +1,0 @@
-def result(query, **kwargs):
-    return [{"this": "plugin loaded"}]
