--- tests/query/mock.mock
@@ -1,4 +1,0 @@
----
-foo: bar
----
-foobar
